Rename modules dir
