+++ src/modules/icons.js
